Bio-inspired Computing/Neuroevolution/Driving Simulation/requirements.txt
graphviz==0.21
neat-python==1.1.0
numpy==2.2.6
opencv-python==4.12.0.88
pygame==2.6.1